Remove dependency on Configuration Manager for .NET Standard
